pyproject.toml
[project]
name = "work-at-codevance"
version = "0.1.0"
description = "Teste técnico pra a Codevance"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Victor Magueta", email = "solermvictor@gmail.com" }
]
dependencies = [
    "asgiref",
    "Django",
    "celery",
    "redis",
    "djangorestframework",
    "djangorestframework-simplejwt",
    "django-celery-results",
    "psycopg2-binary",
    "sqlparse",
]

[project.optional-dependencies]
test = [
    "flake8",
    "black",
    "isort",
    "pytest",
]

[tool.flake8]
ignore = ["F401", "E501", "F841"]
exclude = ["migrations", ".venv", "build"]
max-line-length = 79

[tool.black]
line-length = 79
target-version = ["py38", "py310"]
exclude = '''
/(
    \.eggs
  | \.git
  | \.venv
  | _build
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
src_paths = ["payments"]
multi_line_output = 3
line_length = 79
force_grid_wrap = 0
use_parentheses = true
include_trailing_comma = true
